- move StatisticRqDto from common to lib; - return CommonLoggingErrorHandler for kafka, instead CustomLEH; - add CommonSender to map service - minor fixes;

+++ src/main/java/m3/map/services/StuffService.java
@@ -1,11 +1,11 @@
 package m3.map.services;
 
 public interface StuffService {
-    void giveAHummer(Long userId, int count);
+    void giveAHummer(Long userId, Long count);
 
-    void giveALightning(Long userId, int count);
+    void giveALightning(Long userId, Long count);
 
-    void giveAShuffle(Long userId, int count);
+    void giveAShuffle(Long userId, Long count);
 
-    void giveAGold(Long userId, int count);
+    void giveAGold(Long userId, Long count);
 }
